# apps/api/src/openapi.yaml
openapi: 3.1.0
info:
  title: ashim API
  version: 0.9.0
  description: |
    REST API for ashim, a self-hosted image processing platform with 30+ tools.

    ## Authentication

    Most endpoints require authentication via one of two methods:

    1. **Session cookie** — Call `POST /api/auth/login` with username and password. The response includes a `token` field. Pass it as `Authorization: Bearer <token>` on subsequent requests.

    2. **API key** — Generate a key via the Settings UI or `POST /api/v1/api-keys`. Keys are prefixed with `si_`. Pass as `Authorization: Bearer si_...`.

    Endpoints marked with a lock icon require authentication. Admin-only endpoints are noted in their description.
  license:
    name: AGPL-3.0
    url: https://github.com/ashim-hq/ashim/blob/main/LICENSE
servers:
  - url: /
    description: Current instance
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Session token from login or API key (prefixed with si_)
  schemas:
    Error:
      type: object
      properties:
        statusCode:
          type: integer
          example: 400
        error:
          type: string
          example: Bad Request
        message:
          type: string
          example: Invalid image format
    ToolResponse:
      type: object
      properties:
        jobId:
          type: string
          description: Unique job identifier
        downloadUrl:
          type: string
          description: URL to download the processed image
          example: /api/v1/download/abc123/output.png
        originalSize:
          type: integer
          description: Original file size in bytes
        processedSize:
          type: integer
          description: Processed file size in bytes
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/tools/watermark-image:
    post:
      tags: [Tools]
      summary: Image watermark
      description: Add an image watermark overlay.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, watermark, settings]
              properties:
                file:
                  type: string
                  format: binary
                  description: Main image file
                watermark:
                  type: string
                  format: binary
                  description: Watermark overlay image
                settings:
                  type: string
                  description: |
                    JSON string with options:
                    - `position` (string, default "bottom-right") — One of: center, top-left, top-right, bottom-left, bottom-right
                    - `opacity` (number 0-100, default 50) — Watermark opacity percentage
                    - `scale` (number 1-100, default 25) — Watermark size as percentage of base image
      responses:
        "200":
          description: Processed image
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ToolResponse"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Authentication required
...
  /api/v1/tools/ocr:
    post:
      tags: [Tools]
      summary: OCR / text extraction
      description: Extract text from an image using OCR. Runs locally.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, settings]
              properties:
                file:
                  type: string
                  format: binary
                  description: Image file to extract text from
                settings:
                  type: string
                  description: |
                    JSON string with options:
                    - `engine` (string, default "tesseract") — One of: tesseract, paddleocr
                    - `language` (string, default "en") — One of: en, de, fr, es, zh, ja, ko
      responses:
        "200":
          description: Extracted text
          content:
            application/json:
              schema:
                type: object
                properties:
                  text:
                    type: string
                    description: Extracted text content
                  engine:
                    type: string
                  confidence:
                    type: number
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Authentication required
  /api/v1/tools/info:
    post:
      tags: [Tools]
      summary: Image info
      description: Get detailed metadata about an image including dimensions, format, color space, and EXIF data.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: Image file to inspect
      responses:
        "200":
          description: Image metadata
          content:
            application/json:
              schema:
                type: object
                properties:
                  width:
                    type: integer
                  height:
                    type: integer
                  format:
                    type: string
                  space:
                    type: string
                  channels:
                    type: integer
                  depth:
                    type: string
                  density:
                    type: integer
                  hasAlpha:
                    type: boolean
                  fileSize:
                    type: integer
                  exif:
                    type: object
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Authentication required
...